refactor: split warning tooltips out of TimetableHeader
